requirements.txt
django
djangorestframework
channels
google-cloud-speech
google-cloud-texttospeech
requests
pydub 